Core

This section of the documentation is dedicated to explaining the purpose and usage core functions used throughout the FeatureExtraction module.

Contents

Core

Contents

Index

Functions

Index

ModelSelection.FeatureExtraction.execute!

ModelSelection.FeatureExtraction.featureextraction!

Functions

ModelSelection.FeatureExtraction.featureextraction! — Method

featureextraction!(
    data::ModelSelection.ModelSelectionData;
    fe_sqr::Union{Symbol,Vector{Symbol},Nothing} = nothing,
    fe_log::Union{Symbol,Vector{Symbol},Nothing} = nothing,
    fe_inv::Union{Symbol,Vector{Symbol},Nothing} = nothing,
    fe_lag::Union{Dict{Symbol,Int64},Nothing} = nothing,
    interaction::Union{Vector{Tuple{Symbol,Symbol}},Nothing} = nothing,
    removemissings::Bool = REMOVEMISSINGS_DEFAULT,
    notify = nothing,
)

Perform feature extraction on the data in a model selection object. It applies different feature extraction techniques to the specified variables and updates the data accordingly.

The available feature extraction options are:

fe_sqr: Square the variables.

fe_log: Apply the natural logarithm to the variables.

fe_inv: Invert the variables (1/x).

fe_lag: Lag the variables by a specified order.

interaction: Create interaction variables from pairs of variables.

removemissings: Indicator whether to remove missing values from the data.

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_sqr::Union{Symbol, Vector{Symbol}, Nothing}: (optional) Single symbol or vector of symbols representing variables to be squared.

fe_log::Union{Symbol, Vector{Symbol}, Nothing}: (optional) Single symbol or vector of symbols representing variables to be transformed with the natural logarithm.

fe_inv::Union{Symbol, Vector{Symbol}, Nothing}: (optional) Single symbol or vector of symbols representing variables to be inverted (1/x).

fe_lag::Union{Dict{Symbol, Int64}, Nothing}: (optional) Dictionary specifying the variables and their lag order.

interaction::Union{Vector{Tuple{Symbol, Symbol}}, Nothing}: (optional) Vector of tuples representing variables for interaction.

removemissings::Bool: (optional) Indicator whether to remove missing values from the data.

notify: (optional) Notification method. Default is NOTIFY_DEFAULT.

Returns

data::ModelSelection.ModelSelectionData: Model selection data object with updated data.

Example

data = ModelSelection.ModelSelectionData(...)

featureextraction!(
    data,
    fe_sqr = :x1,
    fe_log = [:x2, :x3],
    fe_inv = :x4,
    fe_lag = Dict(:x5 => 1),
    interaction = [(x6, x7),
    removemissings = true,
]

ModelSelection.FeatureExtraction.execute! — Method

execute!(
    data::ModelSelection.ModelSelectionData;
    fe_sqr::Union{Symbol,Vector{Symbol},Nothing} = nothing,
    fe_log::Union{Symbol,Vector{Symbol},Nothing} = nothing,
    fe_inv::Union{Symbol,Vector{Symbol},Nothing} = nothing,
    fe_lag::Union{Dict{Symbol,Int64},Nothing} = nothing,
    interaction::Union{Vector{Tuple{Symbol,Symbol}},Nothing} = nothing,
    removemissings::Bool = REMOVEMISSINGS_DEFAULT,
    notify = nothing,
)

Execute feature extraction operations on the data in a model selection object. It performs operations such as squaring variables, applying the natural logarithm, inverting variables, lagging variables, and creating interaction variables.

The available feature extraction options are:

fe_sqr: Square the variables.

fe_log: Apply the natural logarithm to the variables.

fe_inv: Invert the variables (1/x).

fe_lag: Lag the variables by a specified order.

interaction: Create interaction variables from pairs of variables.

removemissings: Indicator whether to remove missing values from the data.

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_sqr::Union{Symbol, Vector{Symbol}, Nothing}: (optional) Single symbol or vector of symbols representing variables to be squared.

fe_log::Union{Symbol, Vector{Symbol}, Nothing}: (optional) Single symbol or vector of symbols representing variables to be transformed with the natural logarithm.

fe_inv::Union{Symbol, Vector{Symbol}, Nothing}: (optional) Single symbol or vector of symbols representing variables to be inverted (1/x).

fe_lag::Union{Dict{Symbol, Int64}, Nothing}: (optional) Dictionary specifying the variables and their lag order.

interaction::Union{Vector{Tuple{Symbol, Symbol}}, Nothing}: (optional) Vector of tuples representing variables for interaction.

removemissings::Bool: (optional) Indicator whether to remove missing values from the data.

notify: (optional) Notification method. Default is NOTIFY_DEFAULT.

Returns

data::ModelSelection.ModelSelectionData: Model selection data object with updated data.

Example

data = ModelSelection.ModelSelectionData(...)

execute!(
    data,
    fe_sqr = :x1,
    fe_log = [:x2, :x3],
    fe_inv = :x4,
    fe_lag = Dict(:x5 => 1),
    interaction = [(x6, x7),
    removemissings = true,
]