Utils

This section of the documentation is dedicated to explaining the purpose and usage of the utils function used throughout the FeatureExtraction module.

Contents

Utils

Contents

Index

Functions

Index

ModelSelection.FeatureExtraction.data_add_fe_inv!

ModelSelection.FeatureExtraction.data_add_fe_lag!

ModelSelection.FeatureExtraction.data_add_fe_log!

ModelSelection.FeatureExtraction.data_add_fe_sqr!

ModelSelection.FeatureExtraction.data_add_fe_vars!

ModelSelection.FeatureExtraction.data_add_interaction!

ModelSelection.FeatureExtraction.parse_fe_variables

Functions

ModelSelection.FeatureExtraction.parse_fe_variables — Method

parse_fe_variables(
    fe_vars::Union{Symbol,Vector{Symbol},Dict{Symbol,Int64},Vector{Tuple{Symbol, Symbol}}},
    expvars::Vector{Symbol};
    depvar::Union{Symbol,Nothing} = nothing,
)

This function parses the fixed effect variables specified in fe_vars and returns them as a vector of symbols. It supports different input formats, including single symbols, vectors of symbols, dictionaries with symbols as keys and integers as values, or vectors of tuples with two symbols representing variable combinations. The function also checks the validity of the parsed variables based on the provided expvars (explanatory variables) and depvar (dependent variable). If the parsed variables are empty or not found in the expvars list, the function returns nothing.

Parameters

fe_vars::Union{Symbol, Vector{Symbol}, Dict{Symbol, Int64}, Vector{Tuple{Symbol,Symbol}}}: Fixed effect variables specified as a symbol, vector of symbols, dictionary with symbols as keys and integers as values, or vector of tuples with two symbols.

expvars::Vector{Symbol}: Vector of explanatory variables.

depvar::Union{Symbol, Nothing} = nothing: Dependent variable symbol.

Returns

Vector{Symbol}: If fe_vars is a Symbol or a Vector{Symbol}, a vector of symbols representing the parsed fixed effects variables.

Dict{Symbol,Int64}: If fe_vars is a Dict{Symbol,Int64}, a dict of symbols representing the parsed fixed effects variables with an int related with.

Vector{Tuple{Symbol,Symbol}}: If fe_vars is a Vector{Tuple{Symbol,Symbol}}, a vector of tuple of symbols representing the parsed fixed effects variables tuples.

Nothing: if fe_vars are not in expvars nor depvar.

Example

expvars = [:x1, :x2, :x3]
fe_vars = [:x1, :x3]

vars = parse_fe_variables(fe_vars, expvars; depvar=:y)
# vars: [:x1, :x3]

expvars = [:x1, :x2, :x3]
fe_vars = Dict(:x1 => 1, :x3 => 2)

vars = parse_fe_variables(fe_vars, expvars; depvar=:y)
# vars: Dict(:x1 => 1, :x3 => 2)

expvars = [:x1, :x2, :x3]
fe_vars = [(:x1, :x2), (:x1, :x3)]

vars = parse_fe_variables(fe_vars, expvars; depvar=:y)
# vars: [(:x1, :x2), (:x1, :x3)]

ModelSelection.FeatureExtraction.data_add_fe_vars! — Method

data_add_fe_vars!(
    data::ModelSelection.ModelSelectionData,
    fe_vars::Union{Symbol,Vector{Symbol},Dict{Symbol,Int64}},
    postfix::String,
    func,
)

This function adds fixed effect variables to the data in a model selection object. The fixed effect variables can be specified as a symbol, a vector of symbols, or a dictionary with symbols as keys and integers as values. The postfix argument is a string that will be appended to the variable names, and func is a function that will be applied to the fixed effect variables. If a fixed effect variable already exists in the data.expvars, the function will update the corresponding column in data.expvars_data using the func. Otherwise, the function will create a new column in data.expvars_data and append the variable with the postfix to data.expvars.

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_vars::Union{Symbol, Vector{Symbol}, Dict{Symbol, Int64}}: Fixed effect variables specified as a symbol, vector of symbols, or dictionary with symbols as keys and integers as values.

postfix::String: Postfix string to append to the variable names.

func: Function to apply to the fixed effect variables.

Example

data = ModelSelectionData(...)

fe_vars = [:x2, :x3]
postfix = "_fe"
func(data) = log.(data)

data_add_fe_vars!(data, fe_vars, postfix, func)

ModelSelection.FeatureExtraction.data_add_fe_sqr! — Method

data_add_fe_sqr!(
    data::ModelSelection.ModelSelectionData,
    fe_vars::Union{Symbol,Vector{Symbol}},
)

Add squared fixed effect variables to the data in a model selection object. The fixed effect variables can be specified as a symbol or a vector of symbols. The function applies a square transformation to the fixed effect variables and appends them to data.expvars with the postfix "_sqrt".

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_vars::Union{Symbol, Vector{Symbol}}: Fixed effect variables specified as a symbol or vector of symbols.

Example

data = ModelSelectionData(...)

fe_vars = [:x1, :x2]
data_add_fe_sqr!(data, fe_vars)

ModelSelection.FeatureExtraction.data_add_fe_log! — Method

data_add_fe_log!(
    data::ModelSelection.ModelSelectionData,
    fe_vars::Union{Symbol,Vector{Symbol}},
)

Add logarithmic transformation of fixed effect variables to the data in a model selection object. The fixed effect variables can be specified as a symbol or a vector of symbols. The function applies a logarithmic transformation (log.) to the fixed effect variables and appends them to data.expvars with the postfix "_log".

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_vars::Union{Symbol, Vector{Symbol}}: Fixed effect variables specified as a symbol or vector of symbols.

Errors

ArgumentError(LOG_FUNCTION_ERROR): When negative values are provided.

Example

data = ModelSelectionData(...)

fe_vars = [:x1, :x2]
data_add_fe_log!(data, fe_vars)

ModelSelection.FeatureExtraction.data_add_fe_inv! — Method

data_add_fe_inv!(
    data::ModelSelection.ModelSelectionData,
    fe_vars::Union{Symbol,Vector{Symbol}},
)

Add inverse transformation of fixed effect variables to the data in a model selection object. The fixed effect variables can be specified as a symbol or a vector of symbols. The function applies an inverse transformation (1 ./ data) to the fixed effect variables and appends them to data.expvars with the postfix "_inv".

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_vars::Union{Symbol, Vector{Symbol}}: Fixed effect variables specified as a symbol or vector of symbols.

Example

data = ModelSelectionData(...)

fe_vars = [:x1, :x2]
data_add_fe_inv!(data, fe_vars)

ModelSelection.FeatureExtraction.data_add_fe_lag! — Method

data_add_fe_lag!(
    data::ModelSelection.ModelSelectionData,
    fe_vars::Dict{Symbol,Int64},
)

Add lagged variables of fixed effect variables to the data in a model selection object. The function iterates over the fixed effect variables in fe_vars and generates the lagged variables for each variable based on the specified number of lags. The lagged variables are appended to data.expvars with the postfix "lag" and added to `data.expvarsdataas new columns. The lagged variables are computed for each unique value indata.paneldataif panel data is available. If a fixed effect variable is already present indata.expvars, the lagged values are overwritten indata.expvarsdata. If a fixed effect variable is not present indata.expvars, the lagged values are stored in a separate matrixvardataand later appended todata.expvarsdata`.

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

fe_vars::Dict{Symbol, Int64}: Fixed effect variables and the number of lags specified as a dictionary.

Example

data = ModelSelectionData(...)

fe_vars = Dict(:x1 => 1, :x2 => 2)
data_add_fe_lag!(data, fe_vars)

ModelSelection.FeatureExtraction.data_add_interaction! — Method

data_add_interaction!(
    data::ModelSelection.ModelSelectionData,
    interaction::Vector{Tuple{Symbol,Symbol}},
)

Add interaction variables to the data in a model selection object. The interaction variables are formed by multiplying the values of the specified variables in each tuple. The function applies a logarithmic transformation (log.) to the fixed effect variables and appends them to data.expvars with the postfix "_log".

Parameters

data::ModelSelection.ModelSelectionData: Model selection data object.

interaction::Vector{Tuple{Symbol, Symbol}}: Vector of tuples representing the variables for interaction.

Example

data = ModelSelectionData(...)

interaction = [(:x1, :x2), (:x1, :x3)]
data_add_interaction!(data, interaction)