Core

This section of the documentation is dedicated to explaining the purpose and usage core functions used throughout the Preprocessing module.

Contents

Core

Contents

Index

Functions

Index

ModelSelection.Preprocessing.input

ModelSelection.Preprocessing.input

Functions

Missing docstring.

Missing docstring for Preprocessing.input( equation::String, data::Union{ Array{Float64}, Array{Float32}, Array{Union{Float64,Missing}}, Array{Union{Float32,Missing}}, Tuple, DataFrame, }; datanames::Union{Array{Symbol},Nothing} = nothing, method::Symbol = METHOD_DEFAULT, intercept::Bool = INTERCEPT_DEFAULT, fixedvariables::Union{Symbol,Array{Symbol},Nothing} = FIXED_VARIABLES_DEFAULT, panel::Union{Symbol,Nothing} = PANEL_DEFAULT, time::Union{Symbol,Nothing} = TIME_DEFAULT, seasonaladjustment::Union{Dict{Symbol,Int64},Nothing} = SEASONALADJUSTMENT_DEFAULT, removeoutliers::Bool = REMOVEOUTLIERS_DEFAULT, removemissings::Bool = REMOVEMISSINGS_DEFAULT, notify = NOTIFY_DEFAULT, ). Check Documenter's build log for details.

ModelSelection.Preprocessing.input — Method

input(
    equation::Vector{String},
    data::Union{
        Array{Float64},
        Array{Float32},
        Array{Union{Float64,Missing}},
        Array{Union{Float32,Missing}},
        Tuple,
        DataFrame,
    };
    datanames::Union{Array{Symbol},Nothing} = nothing,
    intercept::Bool = INTERCEPT_DEFAULT,
    fixedvariables::Union{Symbol,Array{Symbol},Nothing} = FIXED_VARIABLES_DEFAULT,
    panel::Union{Symbol,Nothing} = PANEL_DEFAULT,
    time::Union{Symbol,Nothing} = TIME_DEFAULT,
    seasonaladjustment::Union{Dict{Symbol,Int64},Nothing} = SEASONALADJUSTMENT_DEFAULT,
    removeoutliers::Bool = REMOVEOUTLIERS_DEFAULT,
    removemissings::Bool = REMOVEMISSINGS_DEFAULT,
    notify = NOTIFY_DEFAULT,
)

Converts the equation as vector of strings to a vector of symbols, parsing the wildcards and call other functions to perform data preprocessing.

Parameters

equation::Array{String}: The equation in string format.

data::Union{Array{Float32}, Array{Float64}, Array{Union{Float32,Missing}}, Array{Union{Float64,Missing}}, DataFrame, Tuple}: The input data.

datanames::Union{Array{Symbol},Nothing}: (optional) Names of the variables in the data.

intercept::Bool: (optional) Whether to include an intercept in the model. Default is INTERCEPT_DEFAULT.

fixedvariables::Union{Symbol,Array{Symbol},Nothing}: (optional) Fixed variables to include in the model. Default is FIXED_VARIABLES_DEFAULT.

panel::Union{Symbol,Nothing}: (optional) Panel variable for panel data. Default is PANEL_DEFAULT.

time::Union{Symbol,Nothing}: (optional) Time variable for time series data. Default is TIME_DEFAULT.

seasonaladjustment::Union{Dict{Symbol,Int64},Nothing}: (optional) Seasonal adjustment parameters. Default is SEASONALADJUSTMENT_DEFAULT.

removeoutliers::Bool: (optional) Whether to remove outliers from the data. Default is REMOVEOUTLIERS_DEFAULT.

removemissings::Bool: (optional) Whether to remove missing values from the data. Default is REMOVEMISSINGS_DEFAULT.

notify: (optional) Notification method. Default is NOTIFY_DEFAULT.

Equation format

The equation parameter can be in the following formats:

# Vector of strings
equation = ["y", "x1", "x2" "x3"]

# Using wildcards
equation = ["y", "*"]
equation = ["y", "x*"]
equation = ["y", "x1", "z*"]

Returns

modelselection_data: The resulting model selection data.

Errors

ArgumentError(DATANAMES_REQUIRED): When datanames is not provided.

ArgumentError(VARIABLES_NOT_DEFINED): When equation contains undefined variables.

Example

Warning

TODO: seasonaladjustment parameter missing.

equation = ["y", "x1", "x2"]
data = [1.0 1.0 23 1.0 2.0 3.0; 1.0 2.0 33 4.0 5.0 6.0; 2.0 1.0 44 7.0 8.0 9.0]
datanames = [:panel, :time, :y, :x1, :x2, :x3]
job_notify(message::String, data::Union{Any,Nothing} = nothing) = println(message, data)

model = input(
    equation,
    data,
    datanames = datanames,
    method = :fast,
    intercept = true,
    fixedvariables = :x3,
    panel = :panel,
    time = :time,
    removeoutliers = true,
    removemissings = true,
    notify = job_notify,
)
# model: ModelSelectionData(
    equation=[:y, :x1, :x2],
    depvar=:y,
    expvars=[:x1, :x2, :_cons],
    fixedvariables=[:x3],
    panel=:panel,
    time=:time,
    intercept=true,
    datatype=Float32,
    method=:fast,
    nobs=3,
    # ...
)

ModelSelection.Preprocessing.input — Method

input(
    equation::Vector{Symbol},
    data::Union{
        Array{Float64},
        Array{Float32},
        Array{Union{Float64,Missing}},
        Array{Union{Float32,Missing}},
        Tuple,
        DataFrame,
    };
    datanames::Union{Array{Symbol},Nothing} = nothing,
    method::Symbol = METHOD_DEFAULT,
    intercept::Bool = INTERCEPT_DEFAULT,
    fixedvariables::Union{Symbol,Array{Symbol},Nothing} = FIXED_VARIABLES_DEFAULT,
    panel::Union{Symbol,Nothing} = PANEL_DEFAULT,
    time::Union{Symbol,Nothing} = TIME_DEFAULT,
    seasonaladjustment::Union{Dict{Symbol,Int64},Nothing} = SEASONALADJUSTMENT_DEFAULT,
    removeoutliers::Bool = REMOVEOUTLIERS_DEFAULT,
    removemissings::Bool = REMOVEMISSINGS_DEFAULT,
    notify = NOTIFY_DEFAULT,
)

Process different values that are required for data process and validates that the data and the parameters are valid. Converts the data to the selected type and call excecute function to perform data preprocessing.

Parameters

equation::Array{String}: The equation in string format.

data::Union{Array{Float32}, Array{Float64}, Array{Union{Float32,Missing}}, Array{Union{Float64,Missing}}, DataFrame, Tuple}: The input data.

datanames::Union{Array{Symbol},Nothing}: (optional) Names of the variables in the data.

method::Symbol: (optional) The method to use for format the data. Default is METHOD_DEFAULT.

intercept::Bool: (optional) Whether to include an intercept in the model. Default is INTERCEPT_DEFAULT.

fixedvariables::Union{Symbol,Array{Symbol},Nothing}: (optional) Fixed variables to include in the model. Default is FIXED_VARIABLES_DEFAULT.

panel::Union{Symbol,Nothing}: (optional) Panel variable for panel data. Default is PANEL_DEFAULT.

time::Union{Symbol,Nothing}: (optional) Time variable for time series data. Default is TIME_DEFAULT.

seasonaladjustment::Union{Dict{Symbol,Int64},Nothing}: (optional) Seasonal adjustment parameters. Default is SEASONALADJUSTMENT_DEFAULT.

removeoutliers::Bool: (optional) Whether to remove outliers from the data. Default is REMOVEOUTLIERS_DEFAULT.

removemissings::Bool: (optional) Whether to remove missing values from the data. Default is REMOVEMISSINGS_DEFAULT.

notify: (optional) Notification method. Default is NOTIFY_DEFAULT.

Returns

modelselection_data: The resulting model selection data.

Errors

ArgumentError(DATANAMES_REQUIRED): When datanames is not provided.

ArgumentError(INVALID_METHOD): When the specified method is invalid.

ArgumentError(SELECTED_VARIABLES_DOES_NOT_EXISTS): When one or more variables in equation are not present in datanames.

ArgumentError(SELECTED_FIXED_VARIABLES_DOES_NOT_EXISTS): When one or more fixed variables in fixedvariables are not present in datanames.

ArgumentError(SELECTED_FIXED_VARIABLES_IN_EQUATION): When one or more fixed variables in fixedvariables are also present in equation.

ArgumentError(TIME_VARIABLE_INEXISTENT): When the specified time variable does not exist in datanames.

ArgumentError(PANEL_VARIABLE_INEXISTENT): When the specified panel variable does not exist in datanames.

Example

Warning

TODO: seasonaladjustment parameter missing.

equation = [:y, :x1, :x2]
data = [1.0 1.0 23 1.0 2.0 3.0; 1.0 2.0 33 4.0 5.0 6.0; 2.0 1.0 44 7.0 8.0 9.0]
datanames = [:panel, :time, :y, :x1, :x2, :x3]
job_notify(message::String, data::Union{Any,Nothing} = nothing) = println(message, data)

model = input(
    equation,
    data,
    datanames = datanames,
    method = :fast,
    intercept = true,
    fixedvariables = :x3,
    panel = :panel,
    time = :time,
    removeoutliers = true,
    removemissings = true,
    notify = job_notify,
)
# model: ModelSelectionData(
    equation=[:y, :x1, :x2],
    depvar=:y,
    expvars=[:x1, :x2, :_cons],
    fixedvariables=[:x3],
    panel=:panel,
    time=:time,
    intercept=true,
    datatype=Float32,
    method=:fast,
    nobs=3,
    # ...
)

Missing docstring.

Missing docstring for Preprocessing.execute( equation::Vector{Symbol}, data::Union{ Array{Float64}, Array{Float32}, Array{Union{Float64,Missing}}, Array{Union{Float32,Missing}}, }, datanames::Vector{Symbol}, method::Symbol, intercept::Bool, datatype::DataType; fixedvariables::Union{Symbol,Array{Symbol},Nothing} = FIXED_VARIABLES_DEFAULT, panel::Union{Symbol,Nothing} = PANEL_DEFAULT, time::Union{Symbol,Nothing} = TIME_DEFAULT, seasonaladjustment::Union{Dict{Symbol,Int64},Nothing} = SEASONALADJUSTMENT_DEFAULT, removeoutliers::Bool = REMOVEOUTLIERS_DEFAULT, removemissings::Bool = REMOVEMISSINGS_DEFAULT, notify = NOTIFY_DEFAULT, ). Check Documenter's build log for details.